review startup_ keyword and add startup_class_ keyword
The startup_class_ keyword works in the same way as the previous startup_ keyword. In other words, the class name is passed as the parameter (the class must contain the static main member method)

The startup_ keyword now uses the class's main method or any other method representing the program's entry point.